Bugfix for named pipe async hang. On Vista under vmware, namedpipe-2.0 would hang as the message size crossed the 4096 byte boundary.
The problem was that when the first > 4096 message was received,
Tcl channel core would read 4096 bytes. The namedpipe driver
would then queue the thread pool was async read and then set
buffer state to IO_PENDING. However, because data was already
present in the OS, the thread pool would run immediately but not queue
the named pipe event handler because the buffer state was not IO_PENDING.
Fixed by changing state to IO_PENDING before initiating the read.